WT-7619 Add a new optimization to skip pages in cursor traversal when all entries on the page are deleted (#6784)
This change is introducing a new optimizaion that determines if all records on the page have been deleted and all the tombstones are visible to our transaction. If so, we can avoid reading the records on the page and move to the next page. We base this decision on the aggregate stop point added to the page during the last reconciliation. This will help in optimizing cursor traversals for cases where a range of keys have been deleted but can not be removed permanently due to pinned timestamps or long running transactions.