WT-7787 Don't read pages for checkpoint cleanup when the cache is in aggressive mode
When a checkpoint is tasked with garbage collection, it has to
clean up the older content from a file that is now obsolete. To
do so, the checkpoint-cleanup process can effectively visit and
load all of the internal pages in a file. Doing so when the cache
has already reached an aggressive state can be counterproductive.
Not only reading of the pages into the cache for the cleanup can
be slower, but the checkpoint itself could hold eviction in the
aggressive state longer.

When the eviction is set to aggressive, this change allows
checkpointing to skip loading the internal pages that are not
already in cache. This will help a faster checkpoint, and a higher
chance for the cache to get back to normal.

Co-authored-by: Haribabu Kommi <haribabu.kommi@mongodb.com>
Co-authored-by: Sulabh Mahajan <sulabh.mahajan@mongodb.com>