WT-7900 format test program column-store mode doesn't correctly insert new records #6845
Clear the list of inserted records between individual runs of operations (inside a larger, single run): the numbers left in the list can conceivably overlap leading to races, or more likely, the list is entirely filled between runs and an insert in a subsequent run would fail because the count says there are empty slots in the table, and there aren't.

Increment the pointer into the table of inserted records, otherwise we'll never resolve any inserted record that isn't in the first slot of the table. This means we haven't been testing inserts of more than 256 records per thread (and only testing that lightly, this is a significant change).