Change the python test to check for atleast 1 skipped page.
How many pages we skip is not predictable once we start to skip the
internal pages. We can only skip the pages we visit, and if we skip
internal pages, we do not visit the whole subtrees. We should base the
python test checking if we skip pages, to check that we skip atleast one
page.