Accompanying test changes:
1. Enable skipped, disabled, or accidentally masked variable-length
column-store scenarios in these Python tests:

   durable_rollback_to_stable
   durable_ts01
   hs{01,02,06,09,11-13,18}
   prepare{14-16}
   rollback_to_stable{01-08,10-14,18,19,21}
   timestamp{04,06}
   txn07

Also fix prepare14 to not use strings as column-store keys.

2. Add variable-length column store scenarios to these Python tests:

   assert*
   hs{20-25}
   prepare_conflict
   prepare_hs{01,03-05}
   prepare{05-13}
   rollback_to_stable{09,20}
   timestamp{03,05,07}
   txn{03,06,08,13-15,18-20,22-26}

3. Enable the fixed-length column store cases in txn07 and add such
cases to txn{03,06}: these work despite the known issues with
fixed-length column stores, so it's better to enable them so we find
out if they regress further in the future.

4. Add new tests hs{26,27} that exercise the interaction of
column-store RLE cells with timestamps, prepared updates, and the
history store.

5. In hs{12,13}, use eviction cursors instead of trying to stuff the
cache to force eviction.