WT-7830 Migrate the python setup scripts to use cmake (#6827)
Migrate the python pip setup script to build with CMake. It still requires the extension libraries installed, including snappy, zlib, and zstd. If they are not in the default location. You need to configure their location with the CMAKE_LIBRARY_PATH environment variable. In addition, the script automatically installs cmake and ninja for you so there is no need to install them in advance.

Co-authored-by: Alison Felizzi <alison.felizzi@mongodb.com>