Put back this code:
     - Remove code from __wt_bt_col_var_cursor_walk_txn_read that checked
    for an insert update to pass to __wt_txn_read; the insert list has
    already been checked and checking it again is at least useless if not
    actually wrong.

since it is apparently not useless and removing it broke
rollback_to_stable14.

This issue is at least sometimes masked by the optimizer with -O3 and
the breakage only surfaces with -O0 or with the memcmp UB fixes in
rec_col.c. Which is confusing and I don't claim to understand why.