WT-7352 Fix test_hs01 conflict between concurrent operations in cursor modify. (#6855)
Reproduced the error on my machine, however, I do not think that reduced cache is needed for the test functionality. As with the large_modify() (where the error occurs), we are not testing eviction but checking for the modified content in the history store after recovery. Hence, increasing cache to 200MB. Additionally, committing each modification, instead of a huge chunk of modification so to avoid cache pressure. Did not get the errors after making the above changes

Moreover, added statistics to test whether or not data is being inserted into the history store. Also, improved comments.