Don't release clock_lock until after committing.
Otherwise the checkpointer can update the stable timestamp in between
and we end up committing behind it, which then crashes. Only happens
rarely.

Also fixes a problem where transactions can apparently commit out of
order, resulting in assertions about invalid time windows while
handling the history store. I'm not sure if that's expected (that is,
without this change the test is wrong) or if it reflects an actual
problem that this change masks.

To reproduce it reliably, one can revert this change and then also
change the worker thread call to __wt_random_init_seed to just
__wt_random_init; this makes every thread use the same random seed, so
they pound on the same keys at once and readily conflict with one
another.