WT-7718: Renamed 'build_cmake' directory to 'cmake'
We have been using directories prefixed with name 'build_' to
indicate that it's the directory to build inside of. The
'build_cmake' directory contains CMake source code for the build
system and should generally not be used as a build directory.
This could be problematic as building inside 'build_cmake' could
clobber and delete sources that are needed to execute the build.

Renamed 'build_cmake' directory to 'cmake' avoiding potential
confusion.