WT-7947 Created PYTHON3_REQUIRED_VERSION Config
If a users system has multiple python installations, when building
the Python API, CMake will preference using the highest python
version found in the system path. This may be un-ideal if the user
wants to preference a specific python version to build the API
with e.g. ignore another installation.
This change adds the ability to pass a preferred Python version
through a CMake config option. If unset, behaviour will default
to original aforementioned functionality.