Removed DEFAULT string check in config_string
Removed the check in 'config_string' requiring we pass a non-empty
string as a DEFAULT value. Having an empty string can be a valid
input to indicate a 'falsy' value e.g. the config can be visible
since it passes DEPENDS checks but you want to manually set it to
a false empty string as default.