WT-7947 Allow CMake to take in a specific Python version (#6887)
* Created PYTHON3_REQUIRED_VERSION Config

If a users system has multiple python installations, when building
the Python API, CMake will preference using the highest python
version found in the system path. This may be un-ideal if the user
wants to preference a specific python version to build the API
with e.g. ignore another installation.
This change adds the ability to pass a preferred Python version
through a CMake config option. If unset, behaviour will default
to original aforementioned functionality.

* Removed DEFAULT string check in config_string

Removed the check in 'config_string' requiring we pass a non-empty
string as a DEFAULT value. Having an empty string can be a valid
input to indicate a 'falsy' value e.g. the config can be visible
since it passes DEPENDS checks but you want to manually set it to
a false empty string as default.