WT-7953 Teach s_string to not look inside getopt option lists. #6894
Teach s_string to not look at option lists in __wt_getopt calls.
Purge a bunch of unwanted spelling words that are no longer needed. (This doesn't get rid of all of them, because a few also appear in usage messages and those are not so easily filtered. But it helps.)