Remove rollback_to_stable26.
It looks like the reality of integration testing and the available
test machines is that it won't work reliably, and probably can't be
made to.

It isn't worthwhile making a C test for this; the point was to have
something quick in the Python suite. If the behavior regresses, at
least timestamp_abort and I think others in the C suite will start
running forever; this won't necessarily cause overt failures, but
because they're part of make check someone will notice.