WT-7965 Update connection base write generation number at the end of recovery checkpoint (#6909)
During WiredTiger recovery, all the existing transaction ids are reset
based on the page write generation numbers present on the page after
rollback to stable is successfully finished. To avoid having transction ids
on the pages that are written during recovery, WiredTiger updates the
connection base write generation number at the end of the recovery
and save it in the WiredTiger metadata.