WT-7995 Fix the global visibility that it cannot go beyond checkpoint visibility (#6931)
Fix the global visibility of data that it cannot go beyond the checkpoint visibility
when the checkpoint is running. This is to avoid inconsistent data visibility
for the eviction threads operating on the data store when the checkpoint
has finished checkpointing the data store and before the checkpoint thread
checkpoints the history store.