Reworded 'Concurrent transaction IDs' sentence
'that are concurrently running at the time of the snapshot' ->
'that are concurrently active at the time of the snapshot creation'.
And removed "(i.e. have not been committed)" since if a
transaction has started and rolledback, then that txn id will
also be not part of the concurrent txn id list.