WT-7630 Fix a bug of history store losing an entry
To fix the problem of history store store losing the entry
When the checkpoint is running performing a prepared transaction
rollback, the history store entry is not removed if the data store
is already checkpointed the history store entry is marked with
a tombstone. These kind of history store entries is later removed
by the rollback to stable operation or by the next reconciliation.