WT-8019 VLCS snapshot-isolation search mismatch #6952
__Fix accidentally row-specific problem in __wt_txn_modify_check.

__wt_txn_modify_check  wants to check if an on-disk value exists before
looking at its timestamps. It was checking that cbt->ins is NULL,
which is sufficient for row-store but too enthusiastic for
column-store, leading to acceptance of out-of-order writes and
consequent data corruption.

To fix this, change __wt_read_cell_time_window to report whether it
found anything, and only check the results if it did. Determine this
by checking whether the slot is UINT32_MAX and also whether the slots
array exists at all. For FLCS, return failure instead of the default
timestamp. (Note that __wt_txn_modify_check is the only caller of this
code.)

While here, tidy the interface a bit, and in __wt_txn_modify_check
don't bother checking that the cursor isn't NULL; it's not.

Add a Python test (timestamp24) that tickles the problem.