3. In __rollback_abort_ondisk_kv, also don't unpack the value again.
For columns it's always unpacked in the caller already, so pass that
copy in. This is not the case for rows, but it's unpacked
unconditionally so we can move that code to the caller as well for
symmetry.

As before this also means we don't need to pass cip down.