The "ripcip" argument into the visibility mechanism can just be "rip".
It is only used by the row code. There was one case where it was being
used as "cip", but that code was wrong, and (fortunately, then) also
unreachable. ("ins" is never null for column-store.)

Change it to be a WT_ROW * instead of a void *. It turns out that the
corresponding entry in WT_SAVE_UPD was already a WT_ROW *, so just
rename that (ripcip -> rip) for clarity.

It turns out this also allows removing the unsafe casts in WT_ROW_SLOT
and WT_COL_SLOT.