WT-8041 Rollback to stable unpacks values repeatedly (#6963)
* Rearrange the rollback to stable code to avoid repeated unpackings.

* 1. Remove the page argument of __rollback_ondisk_fixup_key.

It was always being passed as null, and besides, the ref is
sufficient. Also, don't bother checking that the ref isn't null; it
isn't.

* 2. In __rollback_ondisk_fixup_key, don't unpack the value again.

It's already been unpacked in the caller, so pass that down. This also
means we don't need to pass cip.

* 3. In __rollback_abort_ondisk_kv, also don't unpack the value again.

For columns it's always unpacked in the caller already, so pass that
copy in. This is not the case for rows, but it's unpacked
unconditionally so we can move that code to the caller as well for
symmetry.

As before this also means we don't need to pass cip down.

* 4. Unpack the row key only once.

If we unpacked the row key up in __rollback_abort_row_leaf, pass it
down to __rollback_ondisk_fixup_key instead of unpacking it again.

* 5. Rationalize the argument order (keys before values).

No further functional change.