Implement timestamps and history for fixed-length column store.
Notes on changes impacting non-FLCS code:

 - Remove ~all the old FLCS special cases.

 - The reconciliation state now potentially handles two chunks of
data; primary (the main page data) and auxiliary (where FLCS pages
keep their timestamp data). Consequently, __wt_rec_split_init now
takes two size arguments, one for the primary size and one for the
auxiliary size (0 if not FLCS). I have avoided overgeneralizing the
code, with the result that there are more tests for FLCS than I'd like
in rec_write.c, but the tests are needed and I don't think there's
anything to be gained by inventing a different state variable to test
against. It's possible that some of it could be factored out more.

 - Without the FLCS special cases __wt_rec_addr_ts_init became a
vacuous wrapper around WT_TIME_AGGREGATE_INIT_MERGE, so I got rid of
it.

 - Rename __rec_leaf_page_max to __rec_leaf_page_max_slvg (in
rec_write.c), for clarity, because it's only used by salvage
reconciliations.

 - Rename __rec_split_grow to __wt_rec_split_grow (in rec_write.c);
this is the code that reallocs page images and it is no longer
file-static so it can be used explicitly from rec_col.c.

 - G/C the obsolete path in __wt_value_return that FLCS had still been
using. With that change, s_all now wants __wt_value_return to return
void, so update all the call sites. (There are quite a few.)

Some other notes:

 - We now set the cursor slot properly in the FLCS code, or at least
comparably to everyone else. It is always either 0 or -1, since
there's only one slot, but it's now in accordance with everything else
so code that looks at the slot doesn't need to be special-cased.

 - Don't turn tombstones into zeros in col_modify. It shouldn't happen
until reconciliation. Storing more zeros in the history store isn't
just inefficient; it also causes test failures.

 - Add a bunch of spelling words. It would be nice if someone figured
out a way to automatically use the identifiers declared inside a
function as spelling words for its comments.