WT-8057 New test to verify the data after aborted compaction (#6986)
This test verifies that there are no data inconsistencies if compact operation is interrupted by an unclean shutdown. To achieve this, the main process spawns a child process. The child process perform certain operations on two identical tables. The parent process randomly kills the child process and verifies that the data across two tables match after restart and recovery.