WT-8161 Reduce verbosity of CMake Evergreen smoke
Currently when ctest executes the smoke target, it does so with
the verbosity option '-VV'. When parsing the logs for potential
failures, this gets extremely noisy since multiple test outputs
are interleaved with each other. To simplify this, remove the 'VV'
option and replace it with 'output-on-failure'. This ensure all
the output for a failing test will be logged, whilst successful
test output will not be logged.