WT-8193 Wrong corner case in VLCS rollback-to-stable (#7038)
* Fix a corner case of VLCS rollback-to-stable.

It turns out that just the presence of a stable update for one key in
the cell isn't sufficient to skip the whole cell. While it's true that
in a fully timestamped world, if there's a stable update it must be
newer than the cell's on-disk value and therefore the cell's on-disk
value must be stable, this breaks down if the stable update isn't
timestamped. We can skip iterating the cell if _every_ key in it has a
stable update; otherwise we need to check each key.

It's also incorrect to call downstream to __rollback_abort_ondisk_kv
for keys that have a stable update; the downstream code assumes that
no stable update exists and will in some cases (e.g. in-memory
databases) do the wrong thing if one does.

Therefore:
   - have __rollback_abort_insert_list return a count of keys with
     stable updates instead of just a flag;
   - add a __rollback_has_stable_update to report if an already
     processed update chain still has a valid and non-aborted
     (therefore stable) update on it;
   - rework the comment about this logic;
   - skip iterating the cell only as described above (or if the cell
     is deleted, as before);
   - when iterating the keys in the cell, also iterate the cell's
     update list to check for stable updates;
   - add a test that exhibits the failing case.

Note that because there's no way to prevent a page being reconciled
before you're ready, without the changes the test might still
occasionally pass. (Reconciling the page breaks the RLE cell in which
things are supposed to happen and this makes the failure
inaccessible.)

This test fails reliably for me without the fixes, but a more
elaborate one I also wrote that tests a lot of related scenarios does
not. (So I'm not including that one.)

* Use correct test number.