WT-7494 Add Python test to trigger update restore eviction during recovery (#7033)
Introduces a python suite test that enables the debug mode setting
for update restore evict during recovery. The debug mode forces
update restore eviction whenever we evict a page. We want to
perform this in recovery to ensure that all the in-memory images
have the proper write generation number and we don't end up
reading stale transaction ID's stored on the page.