WT-8147 Detect invalid CPP config value definition
When parsing the CPP suite configuration string, detect
and raise sytax errors when an unexpected value is found for a
given key. This change specifically raises an error when
we start parsing a new key without having resolved the value of
the previous key i.e. due to a missing comma ','.

Co-authored-by: Etienne Petrel <etienne.petrel@mongodb.com>