WT-8221 Compare write generation before performing RTS
Due to the use of global checkpoint snapshot as a mechanism to
decide to consider to perform RTS on a table and this global
checkpoint snapshot varies for consecutive restarts. RTS must
consider verifying the write generation number with the last
checkpoint connection level base write generation before deciding
whether the table has more updates than the checkpoint snapshot.