Added verbose helper: __wt_verbose_with_events
Since moving out the event definitions into an enum, it no longer
is valid to treat the definitions as flag/mask values. As such
uses of __wt_verbose' will be limited to one event category per
message. In order to still accomodate the uses of __wt_verbose
that would logically 'OR' multiple events, we've added an extra
verbose helper, __wt_verbose_with_events. This accepting a set
of verbose events as a compound-literal array. The macro will
display a verbose message if at least one event in the set
satisfies the required verbosity level.