Lift verbose events size as parameter
Calculating the size of the verbose events array in the
'__wt_verbose_with_events' is error prone and non-computable at
compile-time for any array not declared as a compound literal i.e
not possible if using a ternary expression. Manually lifted out the
event array size as parameter to avoid incorrectly calculating the
size.