WT-6001 Do not read the page into cache if it needs to be rewritten (#7037)
Remove code to acquire a hazard pointer when checking in-memory pages, instead, hold the WT_REF
lock longer. This might be wrong, it means holding the WT_REF locked longer and blocking
application threads, while the block manager locks and reviews extent lists to decide if a
block needs to be rewritten, but it also means we don't have to worry about colliding with
eviction and it's simpler. I don't have any data one way or the other, go for simplicity.
Fix a bug where we weren't setting WT_PAGE_COMPACTION_WRITE on dirty pages.
Fix a bug where we were checking the WT_REF state without holding the WT_REF locked. It's
unlikely to make a difference, either way, and locking first should ensure better compaction
results when racing with other threads.
* Moved compact_progress from bt_compact.c to block_compact.c because accessing WT_BLOCK in the btree layer is pretty big layering violation. Updated the variable names and addressed review comments

Co-authored-by: Keith Bostic <keith.bostic@mongodb.com>