wt_verbose_multi uses compounded category param
To avoid passing two arguments to wt_verbose_multi in tandem
(one to specify the categories and the other for the size), which
can be error prone, compound these two arguments into a struct.
This simplifying the 'wt_verbose_multi' macro and allowing us to
automatically calculate the size of category set.