WT-8285 Avoid global includes in the CMake project
Cleans up the 'config_lib' helper to avoid appending a libraries
include path to the global CMake project scope. We ideally want
to only use the include path if the library is going to be linked
to a target ('config_lib' is only for finding a library). As such
updated 'config_lib' to export an additional configuration variable
containing the libraries include. This allowing us to more
programmatically use the include paths if we are actually going to
link the library.