Create third-party imported library targets
This change centralizes the associated paths under a single shared
library target to avoid dereferencing the 'HAVE_LIB_X' and
'HAVE_LIB_X_INCLUDES' across the project when building against
third party libraries. This greatly simplifying the dependencies
of a library to a single target, which in turn can be easily linked
across the CMake project. This is just setup for the third party
libraries.