Avoid linking tcmalloc in our test binary
If libwiredtiger is already using tcmalloc, linking tcmalloc again
into the binary is potentially error prone. This would result in
two instances of tcmalloc running. A majority of our tests are
simple enough that we generally don't need tcmalloc running in the
test binaries, particularly if wiredtiger is already using
tcmalloc.