WT-6022 NVRAM cache (#7112)
Added the NVRAM block cache to WiredTiger. It uses Optane NVRAM for caching file blocks as  it reads them from or writes them to disk. The cache does not use the persistent property of NVRAM, viewing it as a volatile extension of DRAM. It can be enabled via new configuration options. 

Co-authored-by: Sasha Fedorova <sasha@mongodb.com>
Co-authored-by: Keith Bostic <keith.bostic@mongodb.com>