WT-8255 Python suite test for 'legacy' verbose configuration uses (#7106)
Added a python suite test to assert that basic uses of the verbose
configuration API work as intended i.e. passing single & multiple
valid and invalid verbose categories. This test is mainly focused
on uses of the interface prior to the introduction of verbosity
levels, ensuring 'legacy'-style uses of the interface are still
supported.