WT-3445 Add multiple tables to format tester. (#7079)
* Rework format to handle and test multiple tables. CONFIG changes should be backward compatible, that is,
old and new style CONFIGs should both work. Multiple tables are configured by default, but not if backward compatibility is specified (the -B) option.

* Remove format's ability to do multiple test runs, it now only does a single configuration and test run per
execution.

* Remove the strict order checking on next/prev cursor operations. It's finicky code that only
runs if there are no delete operations and doesn't provide that much value above the cursor
order checking in the library itself. If you put it back, be cautious that you don't check
returned order outside of the initial VLCS bulk-loaded key range, if an insert aborts, then
you'll see skips in the returned cursor values.

* Btree dictionaries apply to both RS & VLCS.

* Turn on additional VLCS testing.

* Don't print out configurations that don't apply to the data source or base configuration.

* Run RTS on in-memory configurations.

* When I removed format's replay support, I didn't quite get it all, remove the underying file name.
Remove a couple of other paths that are no longer in use.

* Add the MAX_FORMAT_PATH #define for general buffer sizing.

* Fix a bug where updates weren't done in implicit transactions as they should have been (they
would happen if the isolation level was left over at snapshot-isolation, but that could be
random).

* Add NVRAM/DRAM block cache support to format.