Rework the compilation of the libwiredtiger target
This change overhauls the CMake build of the wiredtiger target
such that we can compile both a static and shared wiredtiger in
the same build exection i.e. no longer either a static or shared
build.
The main refactors this introduces:
- A macro to define the libwiredtiger target, having a common set
of targets and properties we want to associated to any given
'libwiredtiger' target. Having this as a macro allows us to
de-duplicate common definitions if wanting to create multiple
versions of libwiredtiger within the same build.
- Respect the new 'WITH_PIC' option. If set, ensure static builds
of wiredtiger are compiled as position independent sources (this
has no effect when compiling a shared library target).
- Enforces that at least one build flavour of WiredTiger is enabled
i.e. requires at least 'ENABLE_STATIC' or 'ENABLE_SHARED' is set.