Enforce libwiredtiger build type for Python API
Given the new changes, we can now detect if the user configured a
PIC build. Prior to this change, we were silently converting
static builds to use PIC, which wasn't very transparent to the
end user. Hence updated the Python API build to detect whether a
static PIC build or shared library build is available. If neither
is satisfied, throw an error.