Namespace wiredtiger ALIAS target
As the 'wiredtiger' library target is now an alias target, we
should namespace the target under 'wt::'. This having the benefit
of letting CMake throw diagnostic errors if 'wt::wiredtiger' is
not a CMake target when using 'target_link_libraries'. Otherwise
'target_link_libraries' would silently convert the it to a name
tag i.e. '-lwiredtiger' (which is incorrect).