Coverity reports the statistics cursors can be used uninitialized because it's not (#7157)
Coverity reports the statistics cursors can be used uninitialized because it's not seeing that the WiredTiger library will return an error in any path where a cursor isn't initialized. Change the statistics functions to use format's underlying cursor-open function and change that function to initialize the returned cursor for clarity and to hopefully avoid these complaints in the future.