WT-8335 Support compiling both a static and shared WiredTiger library in CMake (#7138)
Currently we only support compiling either a static or shared build
of the WiredTiger library in CMake (being a binary option). For 
some uses case, it can be ideal to compile both the static and 
shared artifacts in a single build e.g. for a system installation. 
This change overhauls the CMake build of the wiredtiger target such
that we can compile both a static and shared wiredtiger in the same
build run i.e. no longer limiting the user to either a static or shared build. 
This change brings a bit more feature parity with autoconf (before we 
officially deprecate it).

* CMake configuration option: ENABLE_SHARED
* CMake configuration option: WITH_PIC
* Reworks the compilation of the libwiredtiger target
* CMake install support for two wiredtiger targets
* Namespace wiredtiger ALIAS target