WT-8364 Fix CMake bug for TCMalloc cppsuite builds
When building WiredTiger with TCMalloc in CMake, the cppsuite
crashes from an invalid pointer error during a free()'s. When the
tcmalloc config is enabled, CMake builds the cppsuite without
linking tcmalloc (defaulting to the standard libc allocator),
whilst the WiredTiger library is linked against tcmalloc. This
meaning two different allocators are being used at runtime. This
becomes problematic as it appears the cppsuite is
allocating/freeing/sharing memory between itself and WiredTiger,
thus leading to an invalid pointer free() error. To fix this, update
the cppsuite to also link against TCMalloc (if WiredTiger is using
TCMalloc). This keeping the allocator runtime uniform between the
two artifacts.