WT-8364 Fix CMake bug for TCMalloc cppsuite builds (#7169)
On CMake builds of WiredTiger with TCMalloc enabled, the cppsuite
crashes due to an invalid pointer error when free()'ing memory. 
When the tcmalloc config is enabled, CMake builds the cppsuite without
linking tcmalloc (defaulting to the standard libstdc++ allocator),
whilst the WiredTiger library is linked against tcmalloc. This
meaning two different allocators are being used at runtime (TCMalloc
on the C layer & libstdc++ on the C++ layer). This
becomes problematic as it appears the cppsuite is
allocating/freeing/sharing memory between itself and WiredTiger's 
C layer, thus leading to an invalid pointer free() error. To fix this, update
the cppsuite to also link against TCMalloc (if WiredTiger is using
TCMalloc). This keeping the allocator runtime uniform between the
two artifacts.