Make salvage timestamp aggregation the same as reconciliation.
That is, aggregate in a stable timestamp if there were any stable
rows, not just if there were no rows with timestamps. Not absolutely
sure this is necessary, but it's definitely right (if there are stable
values, they should be represented in the aggregate) and even if it's
not mechanically necessary it might become so sometime in the future.

Add a note that we're going to aggregate in all the timestamps, even
if some of them are for rows that the key namespace logic ends up
deciding to throw out, because that code is downstream. The aggregate
will get updated for that change when the page is reconciled after
salvage, and in the meantime things should be ok.

Also comment up the places where we aggregate in stable timestamps,
both in salvage and in reconciliation, because it's not entirely
obvious why that's necessary, especially in bulk load.