Hacks for the "prepare" test in rts10 to make it more reliable under load.
 - Do an initial checkpoint before starting the background
checkpointing thread. The first checkpoint has to write out all the
initial updates and that takes a while. Doing that first increases the
chances that some of the concurrent writes being done will make it
into the checkpoint written by the background thread. (Usually there
is only one, at least for me.) If none of them do, then RTS does no
work and the assertions at the end of the test fail.

 - Sleep for five seconds instead of two to allow the background
checkpointer to get going.

 - Use twice as many rows for FLCS.

This is a bandaid, not a real fix, but makes it much more likely to
succeed for FLCS without compromising the integrity of the test.