WT-8395 Add a version information to the checkpoint snapshot
Adding a version information flag to identify whether the metadata
have a proper snapshot information or not by comparing the retrieved
version information with the binary version.

With this new flag, all the older versions that don't have this flag
are treated having not proper snapshots and ignored them during rollback
to stable operation.