Refactor CMake Build Mode configs
Refactored the CMake build mode configurations to make it more
modular, maintainable and correct. This primarily does the
following refactors:
* Defines a common function given that the ASan and UBSan definitions
follow a common code pattern. This also makes it easier to extend
with new build mode options.
* Distinguish between C and CXX compiler flags, given there can
be variability between the two (rather applying the same flags
across the two toolchains).
* Allow for build modes to be disabled based on dependencies. This
for example needs to be used to disable UBSan on MSVC compilers
(as thats an unavailable option).
* Added MSan and TSan build modes.