Specify an ext subpath in the cppsuite
Instead of baking the in the full system path to the Snappy extension,
just allow CMake to specify whether the extension lives under a
specific subdirectory (i.e. .libs on autoconf). We lose a lot of
portability if we hardcode a full absolute path in the cppsuite,
i.e. the binary becomes invalid if executed under a different
sysroot.