WT-8258 Implement JSON encoded strings to be produced by the WT_EVENT_HANDLER API (#7148)
Currently, messages produced by the event handler
interface are formatted as “flat”-style strings, prepended
with a list of comma-separated attributes. As a means to
improve the structure of our event handler messages, we
updated the event handler to optionally allow the formatting of
messages as JSON-encoded strings. This introduces a structured
and easy-to-parse message format that can be flexibly extended
to expose new metadata attributes.

Specifically, this commit introduces the following key changes:
- Introducing a JSON connection configuration option, "json_output"
- Parsing the JSON connection configuration option
- Refactoring the __eventv function to be able to use different message formats
- Updating test_verbose0[1,2].py to trigger the new code paths

Co-authored-by: Alison Felizzi <alison.felizzi@mongodb.com>